requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0.1
python-multipart>=0.0.6
requests>=2.31.0
